moving code that is not supported by windows
